fix #103 add nodedata view (#104)
* fix #103 add nodedata view
* break GraphView and DataPanle to components
